+++ server.go
@@ -65,7 +65,8 @@ type Server struct {
 
 	shards map[uint64]*Shard // shards by shard id
 
-	Logger *log.Logger
+	Logger     *log.Logger
+	WriteTrace bool // Detailed logging of write path
 
 	authenticationEnabled bool
 
@@ -1369,6 +1370,11 @@ type Point struct {
 // WriteSeries writes series data to the database.
 // Returns the messaging index the data was written to.
 func (s *Server) WriteSeries(database, retentionPolicy string, points []Point) (uint64, error) {
+	if s.WriteTrace {
+		log.Printf("received write for database '%s', retention policy '%s', with %d points",
+			database, retentionPolicy, len(points))
+	}
+
 	// Make sure every point has at least one field.
 	for _, p := range points {
 		if len(p.Fields) == 0 {
@@ -1391,10 +1397,16 @@ func (s *Server) WriteSeries(database, retentionPolicy string, points []Point) (
 	if err := s.createMeasurementsIfNotExists(database, retentionPolicy, points); err != nil {
 		return 0, err
 	}
+	if s.WriteTrace {
+		log.Printf("measurements and series created on database '%s'", database)
+	}
 
 	// Ensure all the required shard groups exist. TODO: this should be done async.
 	if err := s.createShardGroupsIfNotExists(database, retentionPolicy, points); err != nil {
 		return 0, err
+	}
+	if s.WriteTrace {
+		log.Printf("shard groups created for database '%s'", database)
 	}
 
 	// Build writeRawSeriesMessageType publish commands.
@@ -1420,9 +1432,15 @@ func (s *Server) WriteSeries(database, retentionPolicy string, points []Point) (
 			if err != nil {
 				return err
 			}
+			if s.WriteTrace {
+				log.Printf("shard group located: %v", g)
+			}
 
 			// Find appropriate shard within the shard group.
 			sh := g.ShardBySeriesID(series.ID)
+			if s.WriteTrace {
+				log.Printf("shard located: %v", sh)
+			}
 
 			// Many points are likely to have the same Measurement name. Re-use codecs if possible.
 			var codec *FieldCodec
@@ -1445,6 +1463,9 @@ func (s *Server) WriteSeries(database, retentionPolicy string, points []Point) (
 				shardData[sh.ID] = make([]byte, 0)
 			}
 			shardData[sh.ID] = append(shardData[sh.ID], data...)
+			if s.WriteTrace {
+				log.Printf("data appended to buffer for shard %d", sh.ID)
+			}
 		}
 
 		return nil
@@ -1467,6 +1488,9 @@ func (s *Server) WriteSeries(database, retentionPolicy string, points []Point) (
 		if index > maxIndex {
 			maxIndex = index
 		}
+		if s.WriteTrace {
+			log.Printf("write series message published successfully for topic %d", i)
+		}
 	}
 
 	return maxIndex, err
@@ -1481,9 +1505,15 @@ func (s *Server) applyWriteRawSeries(m *messaging.Message) error {
 	if sh == nil {
 		return ErrShardNotFound
 	}
+	if s.WriteTrace {
+		log.Printf("received write message for application, shard %d", sh.ID)
+	}
 
 	if err := sh.writeSeries(m.Data); err != nil {
 		return err
+	}
+	if s.WriteTrace {
+		log.Printf("write message successfully applied to shard %d", sh.ID)
 	}
 
 	return nil